requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
Flask-Script==2.0.6
psycopg2==2.9.1
pandas==1.2.5
python-dotenv
